@@ -1,150 +1,142 @@
 {
-  "name": "vue-enterprise-boilerplate",
-  "author": {
-    "email": "xxx@gmail.com",
-    "name": "xxx"
-  },
+  "name": "vue-default",
   "version": "0.0.0",
   "private": true,
+  "types": "module",
   "scripts": {
-    "build": "vue-cli-service build",
-    "build:ci": "yarn build --report",
-    "dev": "vue-cli-service serve",
-    "dev:e2e": "cross-env VUE_APP_TEST=e2e vue-cli-service test:e2e --mode=development",
-    "docker": "docker-compose exec vue-enterprise-boilderplate yarn",
-    "docs": "vuepress dev",
+    "dev": "vite",
+    "build": "run-p type-check build-only",
+    "preview": "vite preview",
+    "test": "run-s test:unit test:e2e",
+    "test:unit": "vitest --run",
+    "test:ci": "run-s test:unit:ci test:e2e",
+    "test:unit:ci": "yarn test:unit --coverage --run",
+    "test:unit:file": "vitest related --run ",
+    "test:unit:watch": "yarn test:unit",
+    "test:e2e": "cross-env VITE_TEST_E2E=e2e start-server-and-test preview :4173 'cypress run --e2e'",
+    "test:e2e:dev": "cross-env VITE_TEST_E2E=e2e start-server-and-test 'vite dev --port 4173' :4173 'cypress open --e2e'",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
     "lint": "run-s lint:all:*",
-    "lint:all:eslint": "yarn lint:eslint --ext .js,.vue,.ts,.jsx,.tsx .",
+    "lint:all:eslint": "yarn lint:eslint --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts .",
     "lint:all:markdownlint": "yarn lint:markdownlint \"docs/*.md\" \"*.md\"",
-    "lint:all:prettier": "yarn lint:prettier \"**/*.{ts,js,tsx,jsx,json,css,scss,vue,html,md}\"",
+    "lint:all:prettier": "yarn lint:prettier \"**/*.{ts,js,tsx,jsx,cts,mts,cjs,mjs,json,css,scss,vue,html,md}\"",
     "lint:all:stylelint": "yarn lint:stylelint \"src/**/*.{vue,scss}\"",
     "lint:eslint": "eslint --fix",
     "lint:markdownlint": "markdownlint",
-    "lint:prettier": "prettier --write --loglevel warn",
+    "lint:prettier": "prettier --write --loglevel warn --no-error-on-unmatched-pattern",
     "lint:stylelint": "stylelint --fix",
-    "new": "cross-env HYGEN_TMPLS=generators hygen new",
-    "test:unit": "cross-env VUE_APP_TEST=unit vue-cli-service test:unit",
-    "test:e2e": "cross-env VUE_APP_TEST=e2e vue-cli-service test:e2e --headless",
-    "test": "run-s test:unit test:e2e",
-    "test:ci": "run-s test:unit:ci test:e2e",
-    "test:unit:ci": "yarn test:unit --coverage --ci",
-    "test:unit:file": "yarn test:unit --bail --findRelatedTests",
-    "test:unit:watch": "yarn test:unit --watch --notify --notifyMode change",
-    "inspect": "vue inspect --mode production > webpack.config.prod.js",
-    "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(ts|js|vue|tsx|jsx)\" --locales \"./src/lang/**/*.json\"",
-    "ui": "vue ui",
-    "vetur": "vue-int --input src/components --output vetur --recursive -a aliases.config.js#webpack"
+    "docs:dev": "vuepress dev",
+    "docs:build": "vuepress build",
+    "update:dep": "updates -u",
+    "prepare": "is-ci || husky install"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.35",
-    "@fortawesome/free-solid-svg-icons": "5.15.3",
-    "@fortawesome/vue-fontawesome": "3.0.0-3",
-    "@mdi/js": "^5.9.55",
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
+    "@fortawesome/fontawesome-svg-core": "^6.3.0",
+    "@fortawesome/free-solid-svg-icons": "^6.3.0",
+    "@fortawesome/vue-fontawesome": "^3.0.3",
+    "@intlify/unplugin-vue-i18n": "^0.8.2",
+    "@mdi/js": "^7.1.96",
+    "@unhead/vue": "^1.1.8",
+    "axios": "^1.3.2",
     "crypto-browserify": "^3.12.0",
-    "date-fns": "2.10.0",
-    "flatted": "^3.1.1",
-    "lodash": "4.17.19",
-    "normalize.css": "8.0.1",
-    "nprogress": "0.2.0",
-    "tslib": "2.2.0",
-    "vue": "^3.0.0",
-    "vue-i18n": "^9.0.0",
-    "vue-meta": "3.0.0-alpha.4",
-    "vue-router": "^4.0.8",
-    "vuex": "^4.0.1",
-    "vuex-persistedstate": "^4.0.0-beta.3"
+    "date-fns": "2.29.3",
+    "flatted": "^3.2.7",
+    "lodash": "^4.17.21",
+    "normalize.css": "^8.0.1",
+    "nprogress": "^0.2.0",
+    "pinia": "^2.0.30",
+    "vue": "^3.2.47",
+    "vue-i18n": "^9.2.2",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
-    "@babel/compat-data": "^7.13.11",
-    "@babel/core": "^7.13.10",
-    "@babel/preset-env": "^7.13.10",
-    "@commitlint/cli": "^12.0.1",
-    "@commitlint/config-conventional": "^12.0.1",
-    "@intlify/vue-i18n-loader": "^2.1.0",
-    "@types/faker": "^5.5.2",
-    "@types/jest": "^26.0.20",
-    "@types/lodash": "^4.14.168",
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
+    "@esbuild-plugins/tsconfig-paths": "^0.1.2",
+    "@faker-js/faker": "^7.6.0",
+    "@intlify/eslint-plugin-vue-i18n": "^2.0.0",
+    "@pinia/testing": "^0.0.14",
+    "@rushstack/eslint-patch": "^1.2.0",
+    "@types/jsdom": "^21.1.0",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.13.0",
     "@types/nprogress": "^0.2.0",
-    "@types/webpack-env": "^1.16.0",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "@vue/babel-plugin-jsx": "^1.0.3",
-    "@vue/cli-plugin-babel": "~5.0.0-beta.0",
-    "@vue/cli-plugin-e2e-cypress": "~5.0.0-beta.0",
-    "@vue/cli-plugin-eslint": "~5.0.0-beta.0",
-    "@vue/cli-plugin-typescript": "~5.0.0-beta.0",
-    "@vue/cli-plugin-unit-jest": "~5.0.0-beta.0",
-    "@vue/cli-service": "~5.0.0-beta.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "@vue/test-utils": "^2.0.0-0",
-    "autoprefixer": "^10.2.5",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-plugin-macros": "^3.0.1",
-    "compression-webpack-plugin": "^7.1.2",
-    "cross-env": "7.0.x",
-    "eslint": "^7.22.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-promise": "5.1.0",
-    "eslint-plugin-vue": "^7.8.0",
-    "faker": "^5.5.3",
-    "hygen": "6.0.x",
-    "ignore-loader": "^0.1.2",
-    "imagemin-lint-staged": "0.4.x",
-    "jest-extended": "^0.11.5",
-    "lint-staged": "^10.0.x",
-    "local-cypress": "^1.2.1",
-    "markdownlint-cli": "0.22.x",
-    "msw": "^0.28.1",
-    "npm-run-all": "4.1.x",
-    "postcss": "^8.2.10",
-    "postcss-preset-env": "^6.7.0",
+    "@unhead/addons": "^1.1.8",
+    "@vitejs/plugin-legacy": "^4.0.1",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vitejs/plugin-vue-jsx": "^3.0.0",
+    "@vitest/coverage-c8": "^0.28.4",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "@vue/eslint-config-typescript": "^11.0.2",
+    "@vue/test-utils": "^2.2.10",
+    "@vue/tsconfig": "^0.1.3",
+    "@vuepress/client": "^2.0.0-beta.60",
+    "autoprefixer": "^10.4.13",
+    "cross-env": "^7.0.3",
+    "cypress": "^12.5.1",
+    "eslint": "^8.34.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.9.0",
+    "husky": "^8.0.3",
+    "is-ci": "^3.0.1",
+    "jest-matcher-utils": "^29.4.2",
+    "jest-serializer-vue": "^3.1.0",
+    "jsdom": "^21.1.0",
+    "lint-staged": "^13.1.1",
+    "markdownlint-cli": "^0.33.0",
+    "msw": "^1.0.1",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8",
+    "postcss-html": "^1.5.0",
+    "postcss-preset-env": "^8.0.1",
+    "postcss-scss": "^4.0.6",
     "postinstall-postinstall": "^2.1.0",
-    "prettier": "^2.2.1",
-    "require-context.macro": "^1.2.2",
-    "sass": "^1.32.8",
-    "sass-loader": "^11.0.1",
-    "stylelint": "13.2.x",
-    "stylelint-config-css-modules": "2.2.x",
-    "stylelint-config-prettier": "8.0.x",
-    "stylelint-config-recess-order": "2.0.x",
-    "stylelint-config-standard": "20.0.x",
-    "stylelint-scss": "3.14.x",
-    "ts-jest": "^26.5.4",
-    "typedoc": "^0.20.32",
-    "typescript": "~4.1.5",
+    "prettier": "^2.8.4",
+    "prettier-plugin-organize-imports": "^3.2.2",
+    "sass": "^1.58.1",
+    "start-server-and-test": "^1.15.3",
+    "stylelint": "^15.1.0",
+    "stylelint-config-css-modules": "^4.2.0",
+    "stylelint-config-recess-order": "^3.1.0",
+    "stylelint-config-standard": "^30.0.1",
+    "stylelint-config-standard-scss": "^7.0.0",
+    "stylelint-config-standard-vue": "^1.0.0",
+    "stylelint-scss": "^4.4.0",
+    "terser": "^5.16.4",
+    "typedoc": "^0.23.24",
+    "typescript": "~4.9.5",
+    "typescript-vue-plugin": "^1.0.24",
+    "updates": "^13.2.8",
     "utility-types": "^3.10.0",
-    "vls": "^0.7.2",
-    "vue-cli-plugin-i18n": "~2.1.0",
-    "vue-cli-plugin-windicss": "~0.1.5",
-    "vue-intellisense": "^0.3.0",
-    "vue-jest": "^5.0.0-0",
-    "vuepress": "^2.0.0-beta.9",
-    "webpack-bundle-analyzer": "^4.4.1"
-  },
-  "resolutions": {
-    "symbol-observable": "2.0.3",
-    "string_decoder": "1.3.0",
-    "string_decoder/safe-buffer": "5.2.1",
-    "bn.js": "5.2.0"
+    "vite": "^4.1.1",
+    "vite-plugin-compression2": "^0.7.0",
+    "vite-tsconfig-paths": "^4.0.5",
+    "vitest": "^0.28.4",
+    "vue-tsc": "^1.0.24",
+    "vuepress": "^2.0.0-beta.60"
   },
   "engines": {
-    "node": ">=14.0.0",
+    "node": ">=18.0.0",
     "yarn": ">=1.0.0"
   },
-  "gitHooks": {
-    "commit-msg": "commitlint -e -V ",
-    "pre-commit": "cross-env PRE_COMMIT=true lint-staged"
-  },
-  "vetur": {
-    "tags": "vetur/tags.json",
-    "attributes": "vetur/attributes.json"
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
   },
   "msw": {
     "workerDirectory": "public"
+  },
+  "prettier": {
+    "semi": true,
+    "arrowParens": "always",
+    "singleQuote": true,
+    "bracketSpacing": true,
+    "trailingComma": "all",
+    "vueIndentScriptAndStyle": true
   }
 }
