@@ -1,21 +1,17 @@
-<script lang="ts">
-  import { defineComponent, PropType } from 'vue';
-  import { NavRouteItem } from '@models/nav-route-item';
+<script setup lang="ts">
+  import type { NavRouteItem } from '@models/nav-route-item';
+  import type { PropType } from 'vue';
 
-  export default defineComponent({
-    props: {
-      routes: {
-        type: Array as PropType<NavRouteItem[]>,
-        required: true,
-      },
-    },
-    // Render functions are an alternative to templates
-    methods: {
-      getRouteTitle(route: NavRouteItem) {
-        return typeof route.title === 'function' ? route.title() : route.title;
-      },
+  defineProps({
+    routes: {
+      type: Array as PropType<NavRouteItem[]>,
+      required: true,
     },
   });
+
+  function getRouteTitle(route: NavRouteItem): string {
+    return typeof route.title === 'function' ? route.title() : route.title;
+  }
 </script>
 
 <template>
@@ -34,10 +30,12 @@
 <style lang="scss" module>
   @import '@design';
 
-  .active a {
-    font-weight: 600;
-    color: $color-link-text-active;
-    text-decoration: none;
-    cursor: default;
+  .active {
+    a {
+      font-weight: 600;
+      color: $color-link-text-active;
+      text-decoration: none;
+      cursor: default;
+    }
   }
 </style>
