@@ -1,94 +1,82 @@
-<script lang="ts">
-  import appConfig from '@src/app.config.json';
-  import { defineComponent, provide, ref } from 'vue';
-  import { useActiveMeta, useMeta } from 'vue-meta';
-  import { InjectKey } from '@utils/const';
+<script setup lang="ts">
+  // Normalize default styles across browsers,
+  // https://necolas.github.io/normalize.css/
+  import 'normalize.css';
 
-  export default defineComponent({
-    setup() {
-      /**
-       * Vue-meta use Teleport to apply meta data.
-       * {@link https://github.com/nuxt/vue-meta/blob/next/examples/_static/client-only.html}
-       */
-      useMeta({
-        // see metainfo tag in template > v-slot:base in template below
-        // base: { href: '', target: '_blank' },
-        charset: 'utf8',
-        // see metainfo tag in template> v-slot:title in template below
-        title: appConfig.title,
-        description: appConfig.description,
-        // see metainfo tag in template> v-slot:og(title) in template below
-        og: {
-          title: 'Og Title',
-          description: 'Bla bla',
-          image: [
-            'https://picsum.photos/600/400/?image=80',
-            'https://picsum.photos/600/400/?image=82',
-          ],
-        },
-        // this will teleport to head tag
-        head: [
-          {
-            // this will teleport to head tag: <link rel="..." hre="....">
-            // Remove if you dont use
-            tag: 'link',
-            rel: 'stylesheet',
-            href:
-              'https://fonts.googleapis.com/css2?family=Nunito:ital,wght@0,200;0,300;0,400;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,600;1,700;1,800;1,900&display=swap',
-          },
-        ],
-        // this will teleport to body tag. see metainfo tag in template> v-slot:body in template below.
-        // Still bugged in ver vue-meta v3.0.0-alpha.4.
-        // body: [{ tag: 'script', href: '', rel: '' }],
-      });
-      // Current metadat
-      const metadata = useActiveMeta();
-      // Contain current layout with default is default.vue
-      const currentLayout = ref('default');
-      // Change view layout
-      const changeLayout = (layoutFileName: string) => {
-        if (currentLayout.value !== layoutFileName) {
-          currentLayout.value = layoutFileName;
-        }
-      };
-      // Provide to router/layouts/layout.vue to trigger update layout
-      provide(InjectKey.currentLayout, currentLayout);
-      provide(InjectKey.changeLayout, changeLayout);
-      return {
-        currentLayout,
-        metadata,
-      };
-    },
+  // Style loading bar between pages.
+  // https://github.com/rstacruz/nprogress
+  import { useRootStore } from '@stores';
+  import { useHead, useSeoMeta } from '@unhead/vue';
+  import 'nprogress/nprogress.css';
+  import { storeToRefs } from 'pinia';
+  import { useI18n } from 'vue-i18n';
+
+  const { t } = useI18n();
+
+  /**
+   * use seometa for some popular site like: google, twitter...
+   * or use `useHead` for fully customize meta
+   * {@link https://harlanzw.com/blog/vue-use-head-v1#v100-release}
+   */
+  // this will teleport to head tag
+  useSeoMeta({
+    charset: 'utf8',
+    ogImage: [
+      {
+        url: 'https://picsum.photos/600/400/?image=80',
+        alt: '',
+      },
+      {
+        url: 'https://picsum.photos/600/400/?image=82',
+        alt: '',
+      },
+    ],
+    ogTitle: 'Og title',
+    ogDescription: 'Og description',
   });
+
+  useHead({
+    title: t('app.title'),
+    // Configures dynamic template to customize the page title on an individual page.
+    titleTemplate: (title) => `My App - ${title}`,
+
+    // these tags will teleport into head tag
+    // Each element in the array is mapped to a newly-created <meta> tag, where object properties are mapped to the corresponding attributes.
+    meta: [
+      {
+        name: 'viewport',
+        content: 'width=device-width, initial-scale=1, maximum-scale=1',
+      },
+      { name: 'og:title', content: 'Og title' },
+      { name: 'og:description', content: 'Og description' },
+      { name: 'og:image', content: 'https://picsum.photos/600/400/?image=80' },
+      { name: 'og:image', content: 'https://picsum.photos/600/400/?image=82' },
+      { name: 'description', content: t('app.description') },
+    ],
+    // Each element in the array is mapped to a newly-created <link> tag, where object properties are mapped to the corresponding attributes.
+    link: [
+      {
+        as: 'style',
+        rel: 'stylesheet',
+        href: 'https://fonts.googleapis.com/css2?family=Nunito:ital,wght@0,200;0,300;0,400;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,600;1,700;1,800;1,900&display=swap',
+      },
+    ],
+    // Each element in the array is mapped to a newly-created <style> tag, where object properties are mapped to the corresponding attributes.
+    style: [],
+    // Each element in the array is mapped to a newly-created <script> tag, where object properties are mapped to the corresponding attributes.
+    script: [],
+    // Each element in the array is mapped to a newly-created <noscript> tag, where object properties are mapped to the corresponding attributes.
+    noscript: [],
+    // Sets attributes of the <body> tag. Each object property is mapped to the corresponding attribute.
+    bodyAttrs: {},
+    //Sets attributes of the <html> tag. Each object property is mapped to the corresponding attribute.
+    htmlAttrs: {},
+  });
+  const { currentViewLayout } = storeToRefs(useRootStore());
 </script>
 
 <template>
-  <!--
-    vue-meta use Teleport to apply meta data.
-    https://v3.vuejs.org/guide/teleport.html
-    https://github.com/nuxt/vue-meta/blob/next/examples/_static/client-only.html
-   -->
-  <metainfo>
-    <!-- <template v-slot:base="{ content }">
-      {{ content }}
-    </template> -->
-    <template v-slot:title="{ content, metainfo }">
-      {{ content }} - {{ metainfo.description }}
-    </template>
-    <template v-slot:og(title)="{ content, metainfo, og }">
-      {{ content }} - {{ og.description }} - {{ metainfo.description }}
-    </template>
-    <template v-slot:body="{ metainfo }">
-      <component
-        :is="script.tag"
-        v-for="script in metainfo"
-        :key="script.href"
-        :rel="script.rel"
-        :src="script.href"
-      ></component>
-    </template>
-  </metainfo>
-  <component :is="currentLayout" v-bind="$attrs">
+  <component :is="currentViewLayout" v-bind="$attrs">
     <router-view />
   </component>
 </template>
@@ -98,18 +86,9 @@
   // Allow element/type selectors, because this is global CSS.
   // stylelint-disable selector-max-type, selector-class-pattern
 
-  // Normalize default styles across browsers,
-  // https://necolas.github.io/normalize.css/
-  @import '~normalize.css/normalize.css';
-  // Style loading bar between pages.
-  // https://github.com/rstacruz/nprogress
-  @import '~nprogress/nprogress.css';
-
   // Design variables and utilities from src/design.
   @import '@design';
-  // @import 'tailwindcss/base';
-  // @import 'tailwindcss/components';
-  // @import 'tailwindcss/utilities';
+
   *,
   *::before,
   *::after {
@@ -119,6 +98,7 @@
   body {
     background: $color-body-bg;
   }
+
   #app {
     @extend %typography-small;
   }
