@@ -1,46 +1,88 @@
 <script lang="ts">
-  import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
-  import { library as fontAwesomeIconLibrary } from '@fortawesome/fontawesome-svg-core';
+  export default {
+    inheritAttrs: false,
+  };
+</script>
+
+<script setup lang="ts">
+  // icons source
+  import {
+    library as fontAwesomeIconLibrary,
+    type IconDefinition,
+    type IconPack,
+  } from '@fortawesome/fontawesome-svg-core';
+  import {
+    FontAwesomeIcon,
+    type FontAwesomeIconProps,
+  } from '@fortawesome/vue-fontawesome';
+
+  import { IconSource } from '@src/constraint/const';
   import camelCase from 'lodash/camelCase';
-  import { defineComponent } from 'vue';
-  import * as faSync from '@fortawesome/free-solid-svg-icons/faSync';
-  import * as faUser from '@fortawesome/free-solid-svg-icons/faUser';
+  import { computed, useAttrs, useCssModule, type PropType } from 'vue';
 
   // https://fontawesome.com/icons
-  fontAwesomeIconLibrary.add(faSync.definition, faUser.definition);
-  export default defineComponent({
-    components: {
-      FontAwesomeIcon,
+  const styles = useCssModule();
+  const attrs = useAttrs();
+  const props = defineProps({
+    source: {
+      type: String as PropType<IconSource>,
+      default: IconSource.AWESOME_FONT,
     },
-    inheritAttrs: false,
-    props: {
-      source: {
-        type: String,
-        default: 'font-awesome',
-      },
-      name: {
-        type: String,
-        required: true,
-      },
+    icon: {
+      type: [Object, String] as PropType<IconDefinition | IconPack | string>,
+      required: true,
     },
-    computed: {
-      // Gets a CSS module class, e.g. iconCustomLogo
-      customIconClass(): string {
-        return this.$style[camelCase('icon-custom-' + this.name)];
+    size: {
+      type: [String, Number] as PropType<FontAwesomeIconProps['size'] | number>,
+      default: (rawProps: FontAwesomeIconProps['size'] | number) => {
+        if (typeof rawProps === 'number') {
+          return 16;
+        }
+        return 'sm';
       },
     },
+  });
+  const getSize: any = computed(() => {
+    if (props.source === IconSource.AWESOME_FONT) {
+      return typeof props.size === 'string' ? props.size : 'sm';
+    }
+    if (props.source === IconSource.MDI_FONT) {
+      return (typeof props.size === 'number' ? props.size : 32) + 'px';
+    }
+    return 16;
+  });
+
+  const customIconClass = computed(() => {
+    if (props.source !== IconSource.CUSTOM) return '';
+    return styles[camelCase(`icon-custom-${props.icon}`)];
+  });
+
+  const awesomeIcon = computed(() => {
+    if (typeof props.icon !== 'string') {
+      fontAwesomeIconLibrary.add(props.icon);
+      return props.icon;
+    }
+    return '';
   });
 </script>
 
 <template>
   <FontAwesomeIcon
-    v-if="source === 'font-awesome'"
-    v-bind="$attrs"
-    :icon="name"
+    v-if="source === IconSource.AWESOME_FONT"
+    v-bind="attrs"
+    :icon="awesomeIcon"
+    :size="getSize"
   />
-  <span
-    v-else-if="source === 'custom'"
-    v-bind="$attrs"
-    :class="customIconClass"
-  />
+  <svg
+    v-else-if="source === IconSource.MDI_FONT && typeof icon === 'string'"
+    :width="getSize"
+    :height="getSize"
+  >
+    <path :d="icon" />
+  </svg>
+  <span v-else v-bind="attrs" :class="customIconClass" />
 </template>
+
+<!-- <style module lang="scss"> -->
+<!--   // style -->
+<!-- </style> -->
