@@ -1,43 +1,57 @@
 <script lang="ts">
-  import { defineComponent } from 'vue';
-  export default defineComponent({
-    props: {
-      modelValue: {
-        type: String,
-        default: undefined,
-      },
-      type: {
-        type: String,
-        default: 'text',
-        validator: (val: string) =>
-          [
-            'email',
-            'number',
-            'password',
-            'search',
-            'tel',
-            'text',
-            'url',
-          ].includes(val),
-      },
+  export default {
+    inheritAttrs: false,
+  };
+</script>
+
+<script setup lang="ts">
+  import { computed } from 'vue';
+
+  /**
+   * Manual Setup v-model with extra validator
+   * {@link https://vuejs.org/guide/components/v-model.html}
+   */
+  const props = defineProps({
+    modelValue: {
+      type: String,
     },
-    emits: ['update:modelValue'],
-    computed: {},
+    type: {
+      type: String,
+      default: 'text',
+      validator: (val: string) =>
+        [
+          'email',
+          'number',
+          'password',
+          'search',
+          'tel',
+          'text',
+          'url',
+        ].includes(val),
+    },
+  });
+  const emit = defineEmits<{
+    (e: 'update:modelValue', value?: string): void;
+  }>();
+  // Manual v-model. Better than $event.target.value (not typesafe)
+  // https://vuejs.org/guide/components/v-model.html
+  const value = computed({
+    get() {
+      return props.modelValue;
+    },
+    set(value?: string) {
+      emit('update:modelValue', value);
+    },
   });
 </script>
 
 <template>
-  <input
-    :type="type"
-    :class="$style.input"
-    :value="modelValue"
-    v-bind="$attrs"
-    @input="$emit('update:modelValue', $event.target.value)"
-  />
+  <input :type="type" :class="$style.input" v-bind="$attrs" v-model="value" />
 </template>
 
 <style lang="scss" module>
   @import '@design';
+
   .input {
     @extend %typography-small;
 
