@@ -1,95 +1,104 @@
-<script lang="ts">
-  import { defineComponent, PropType } from 'vue';
-  import { RouteLocationRaw } from 'vue-router';
+<script setup lang="ts">
   import isString from 'lodash/isString';
+  import { onMounted, watchEffect, type PropType } from 'vue';
+  import type { RouteLocationRaw } from 'vue-router';
+  const props = defineProps({
+    allowInsecure: {
+      type: Boolean,
+    },
+    to: {
+      type: [Object, String] as PropType<RouteLocationRaw | string>,
+    },
+    params: {
+      type: Object,
+      default: () => ({}),
+    },
+    name: {
+      type: String,
+    },
+    href: {
+      type: String,
+    },
+    target: {
+      type: String,
+      default: '_blank',
+    },
+  });
 
-  export default defineComponent({
-    props: {
-      allowInsecure: {
-        type: Boolean,
-        default: false,
-      },
-      to: {
-        type: [Object, String] as PropType<RouteLocationRaw>,
-        default: null,
-      },
-      params: {
-        type: Object,
-        default: () => ({}),
-      },
-      name: {
-        type: String,
-        default: undefined,
-      },
-      href: {
-        type: String,
-        default: undefined,
-      },
-      target: {
-        type: String,
-        default: '_blank',
-      },
-    },
-    computed: {},
-    created() {
-      this.validateProps();
-    },
-    methods: {
-      routerLinkTo(route?: Record<string, any>) {
-        if (isString(this.to)) return this.to;
-        return {
-          name: route?.name || this.name,
-          params: route?.params || this.params,
-          ...(this.to || {}),
-        };
-      },
-      validateUrlSecure() {
-        const isInsecureUrl =
-          !this.allowInsecure && !/^(https|mailto|tel):/.test(this.href);
-        if (isInsecureUrl) {
-          return console.warn(
-            `Insecure <BaseLink> href: ${this.href}.\nWhen linking to external sites, always prefer https URLs. If this site does not offer SSL, explicitly add the allow-insecure attribute on <BaseLink>.`
-          );
-        }
-      },
-      validateUrlIsLocal() {
-        const isLocalUrl = !/^\w+:/.test(this.href);
-        if (isLocalUrl) {
-          return console.warn(
-            `Invalid <BaseLink> href: ${this.href}.\nIf you're trying to link to a local URL, provide at least a name or to`
-          );
-        }
-      },
-      validateRouterLinkProp() {
-        const insufficientProp = !this.name && !this.to;
-        if (insufficientProp) {
-          return console.warn(
-            `Invalid <BaseLink> props:\n\n${JSON.stringify(
-              this.$props,
-              null,
-              2
-            )}\n\nEither a \`name\` or \`to\` is required for internal links, or an \`href\` for external links.`
-          );
-        }
-      },
-      // Perform more complex prop validations than is possible
-      // inside individual validator functions for each prop.
-      validateProps() {
-        if (process.env.NODE_ENV === 'production') return;
-        if (this.href) {
-          this.validateUrlSecure();
-          this.validateUrlIsLocal();
-        } else {
-          // Check for insufficient props.
-          this.validateRouterLinkProp();
-        }
-      },
-    },
+  function routerLinkTo(route?: Record<string, any>) {
+    if (isString(props.to)) return props.to;
+    return {
+      name: route?.name || props.name,
+      params: route?.params || props.params,
+      ...(props.to || {}),
+    };
+  }
+
+  function validateUrlSecure() {
+    const isInsecureUrl =
+      props.href &&
+      !props.allowInsecure &&
+      !/^(https|mailto|tel):/.test(props.href);
+
+    if (isInsecureUrl) {
+      return console.warn(
+        `Insecure <BaseLink> href: ${props.href}.\nWhen linking to external sites, always prefer https URLs. If this site does not offer SSL, explicitly add the allow-insecure attribute on <BaseLink>.`,
+      );
+    }
+  }
+
+  function validateUrlIsLocal() {
+    const isLocalUrl = props.href && !/^\w+:/.test(props.href);
+
+    if (isLocalUrl) {
+      return console.warn(
+        `Invalid <BaseLink> href: ${props.href}.\nIf you're trying to link to a local URL, provide at least a name or to`,
+      );
+    }
+  }
+
+  function validateRouterLinkProp() {
+    const insufficientProp = !props.name && !props.to;
+
+    if (insufficientProp) {
+      return console.warn(
+        `Invalid <BaseLink> props:\n\n${JSON.stringify(
+          props,
+          null,
+          2,
+        )}\n\nEither a \`name\` or \`to\` is required for internal links, or an \`href\` for external links.`,
+      );
+    }
+  }
+
+  // Perform more complex prop validations than is possible
+  // inside individual validator functions for each prop.
+  function validateProps() {
+    if (props.href) {
+      validateUrlSecure();
+      validateUrlIsLocal();
+    } else {
+      // Check for insufficient props.
+      validateRouterLinkProp();
+    }
+  }
+
+  watchEffect(() => {
+    validateProps();
+  });
+
+  onMounted(() => {
+    validateProps();
   });
 </script>
 
 <template>
-  <a v-if="!!href" :href="href" :target="target" v-bind="$attrs">
+  <a
+    v-if="!!props.href"
+    :href="props.href"
+    :target="props.target"
+    v-bind="$attrs"
+  >
     <slot />
   </a>
   <RouterLink v-else v-slot="scopedSlot" :to="routerLinkTo()" custom>
