@@ -1,37 +1,33 @@
-<script lang="ts">
-  import {
-    defineComponent,
-    onBeforeUnmount,
-    onMounted,
-    PropType,
-    ref,
-    toRefs,
-  } from 'vue';
+<script lang="ts" setup>
+  import { onBeforeUnmount, onMounted, ref, toRefs, type PropType } from 'vue';
 
-  export default defineComponent({
-    props: {
-      callback: {
-        type: Function as PropType<() => void>,
-        required: true,
-      },
+  const props = defineProps({
+    callback: {
+      type: Function as PropType<() => void>,
+      required: true,
     },
-    setup(props) {
-      const { callback } = toRefs(props);
-      const root = ref<HTMLElement>();
-      const listener = (e) => {
-        if (e.target === root.value || root.value?.contains(e.target)) {
-          return;
-        }
-        callback.value();
-      };
-      onMounted(() => {
-        document.addEventListener('click', listener);
-      });
-      onBeforeUnmount(() => {
-        document.removeEventListener('click', listener);
-      });
-      return { root };
-    },
+  });
+
+  const { callback } = toRefs(props);
+
+  // declare a ref to hold the element reference
+  // the name must match template ref value
+  const root = ref<HTMLElement | null>(null);
+  const listener = (e: MouseEvent) => {
+    if (
+      e.target === root.value ||
+      root.value?.contains(e.target as Node | null)
+    ) {
+      return;
+    }
+    callback.value();
+  };
+
+  onMounted(() => {
+    document.addEventListener('click', listener);
+  });
+  onBeforeUnmount(() => {
+    document.removeEventListener('click', listener);
   });
 </script>
 
