@@ -1,128 +1,126 @@
-import { createRouter, createWebHistory, NavigationGuard } from 'vue-router';
-// Adds a loading bar at the top during page loads.
-import NProgress from 'nprogress';
-import { store } from '@state/store';
-import { reactive, toRefs } from 'vue';
+import { i18nGlobal } from '@plugin';
+import { getLocale, isLocaleSupported, setI18nLanguage } from '@plugin/i18n';
+import { useAuthStore } from '@stores';
 import isArray from 'lodash/isArray';
 import isEmpty from 'lodash/isEmpty';
-import { i18n, loadLocaleMessages, SUPPORT_LOCALES } from '@plugin/i18n';
+import NProgress from 'nprogress';
+import { unref } from 'vue';
+import {
+  createRouter as originalCreateRoute,
+  createWebHistory,
+  type RouteLocationNormalized,
+  type RouteLocationRaw,
+} from 'vue-router';
 import routes from './routes';
 
-const router = createRouter({
-  routes,
-  history: createWebHistory(),
-  scrollBehavior(to, from, savedPosition) {
-    if (savedPosition) {
-      return savedPosition;
-    } else {
-      return { left: 0, right: 0 };
+export async function createRouter() {
+  const router = originalCreateRoute({
+    routes,
+    history: createWebHistory(),
+    scrollBehavior: (_to, _from, savedPosition) => {
+      if (savedPosition) {
+        return savedPosition;
+      } else {
+        return { left: 0, right: 0 };
+      }
+    },
+  });
+
+  // Before each route evaluates...
+  router.beforeEach(async (to, from) => {
+    // If this isn't an initial page load...
+    if (from.name !== null) {
+      // Start the route progress bar.
+      NProgress.start();
     }
-  },
-});
 
-// Before each route evaluates...
-router.beforeEach(async (to, from) => {
-  // If this isn't an initial page load...
-  if (from.name !== null) {
-    // Start the route progress bar.
-    NProgress.start();
-  }
+    // Get global locale
+    const locale = getLocale(i18nGlobal);
 
-  const { locale } = toRefs(reactive(i18n.global));
-  const currentLocale = locale.value;
+    const currentLocale = unref(locale);
+    const localeInParams = getLocaleFromParams(to.params);
+    const isNewLocale = localeInParams !== currentLocale;
 
-  const localeInParams = getLocaleFromParams(to.params);
-
-  const isNewLocale = localeInParams !== currentLocale;
-  if (isNewLocale) {
-    const isNewLocaleSupported = SUPPORT_LOCALES.includes(localeInParams);
-    if (!isNewLocaleSupported && to.name !== '404') {
-      return redirectToFallbackLocale(currentLocale, to.fullPath);
+    if (isNewLocale) {
+      // case forget to add locale or locale not exist
+      if (!isLocaleSupported(localeInParams)) {
+        return redirectToFallbackLocale(currentLocale, to);
+      }
+      // update i18n plugin with new locale
+      await setI18nLanguage(i18nGlobal, localeInParams);
     }
-    if (localeInParams !== undefined)
-      await localeMessageChange(i18n, localeInParams);
-  }
 
-  // Check if auth is required on this route
-  // (including nested routes).
-  const authRequired = to.meta?.authRequired;
+    // Check if auth is required on this route
+    // (including nested routes).
+    const authRequired = to.meta?.authRequired;
 
-  // If auth isn't required for the route, just continue.
-  if (!authRequired) return true;
+    // If auth isn't required for the route, just continue.
+    if (!authRequired) return true;
 
-  // If auth is required and the user is logged in...
-  if (store.getters['auth/loggedIn']) {
-    // Validate the local user token...
-    return store.dispatch('auth/validate').then((validUser) => {
-      // Then continue if the token still represents a valid user,
-      // otherwise redirect to login.
-      return validUser ? true : redirectToLoginPage(to.fullPath);
-    });
-  }
+    // If auth is required and the user is logged in...
+    const authStore = useAuthStore();
 
-  // If auth is required and the user is NOT currently logged in,
-  // redirect to login.
-  return redirectToLoginPage(to.fullPath);
-});
+    if (authStore.loggedIn) {
+      // Validate the local user token...
+      const validUser = await authStore.validate();
 
-router.beforeResolve((to, from) => {
-  // Create a `beforeResolve` hook, which fires whenever
-  // `beforeRouteEnter` and `beforeRouteUpdate` would. This
-  // allows us to ensure data is fetched even when params change,
-  // but the resolved route does not. We put it in `meta` to
-  // indicate that it's a hook we created, rather than part of
-  // Vue Router (yet?).
-  try {
-    return new Promise((resolve) => {
+      return validUser ? true : redirectToLoginPage(to);
+    }
+
+    // If auth is required and the user is NOT currently logged in,
+    // redirect to login.
+    return redirectToLoginPage(to);
+  });
+
+  router.beforeResolve(async (to, from) => {
+    // Create a `beforeResolve` hook, which fires whenever
+    // `beforeRouteEnter` and `beforeRouteUpdate` would. This
+    // allows us to ensure data is fetched even when params change,
+    // but the resolved route does not. We put it in `meta` to
+    // indicate that it's a hook we created, rather than part of
+    // Vue Router (yet?).
+    try {
       if (typeof to.meta?.beforeResolve === 'function') {
-        to.meta.beforeResolve(
-          to,
-          from,
-          (
-            redirectOpts?: ReturnType<
-              Parameters<typeof router.beforeResolve>[0]
-            >
-          ) => {
-            if (!redirectOpts) resolve(true);
-            // If the user chose to redirect...
-            // If redirecting to the same route we're coming from...
-            if (
-              typeof redirectOpts === 'object' &&
-              'name' in redirectOpts &&
-              from.name === redirectOpts.name
-            ) {
-              // Complete the animation of the route progress bar.
-              NProgress.done();
-            }
-            // Complete the redirect.
-            return resolve(redirectOpts);
-          }
-        );
-      } else resolve(true);
-    });
-  } catch (error) {
-    return error;
-  }
-});
+        const redirectTo = await to.meta.beforeResolve(to, from);
 
-// When each route is finished evaluating...
-router.afterEach(() => {
-  // Complete the animation of the route progress bar.
-  NProgress.done();
-});
+        // If the user chose to redirect...
+        // and redirecting to the same route we're coming from...
+        if (
+          typeof redirectTo === 'object' &&
+          'name' in redirectTo &&
+          from.name === redirectTo.name
+        ) {
+          // Complete the animation of the route progress bar.
+          NProgress.done();
+        }
+        // Complete the redirect.
+        return redirectTo;
+      } else return true;
+    } catch (error: any) {
+      // return error;
+      return false;
+    }
+  });
 
-async function localeMessageChange(i18n, locale) {
-  if (!i18n.global.availableLocales.includes(locale)) {
-    await loadLocaleMessages(i18n, locale);
-  }
+  // When each route is finished evaluating...
+  router.afterEach(() => {
+    // Complete the animation of the route progress bar.
+    NProgress.done();
+  });
+  return router;
 }
 
-function getLocaleFromParams(params: Record<string, any>): string {
-  let selectedLocale = (isArray(params.locale)
-    ? !isEmpty(params.locale)
-      ? params.locale[0]
-      : undefined
-    : params.locale
+/**
+ * @param params - Current params in this route. Example: /en/404 -> matched route :locale/404 -> params is {locale: en}
+ * @returns return locale value. Example: en
+ */
+export function getLocaleFromParams(params: Record<string, any>): string {
+  let selectedLocale = (
+    isArray(params.locale)
+      ? !isEmpty(params.locale)
+        ? params.locale[0]
+        : undefined
+      : params.locale
   )?.toString();
 
   if (
@@ -134,18 +132,40 @@ function getLocaleFromParams(params: Record<string, any>): string {
   return selectedLocale;
 }
 
-function redirectToFallbackLocale(
+/**
+ * @param fallbackLocale - redirect to fallback locale. Ex: /not-exist-locale/profile -> /en/profile
+ * @param routeName - redirect from this path to `/:fallbackLocale/originpath`.
+ * @param query - query from current path
+ * @returns route with new locale
+ */
+export function redirectToFallbackLocale(
   fallbackLocale: string,
-  originFullPath: string
-) {
-  return `/${fallbackLocale}${originFullPath}`;
+  to?: RouteLocationNormalized,
+): RouteLocationRaw {
+  // /en/originPath?query
+  return {
+    name: (to?.name as string) || undefined,
+    params: {
+      ...to?.params,
+      locale: fallbackLocale,
+    },
+    query: to?.query,
+    hash: to?.hash,
+
+    // path: newPath,
+    replace: true,
+  };
 }
 
-function redirectToLoginPage(
-  redirectFrom: string
-): ReturnType<NavigationGuard> {
+function redirectToLoginPage(to: RouteLocationNormalized): RouteLocationRaw {
   // Pass the original route to the login component
-  return { name: 'login', query: { redirectFrom } };
+  // /en/login
+  return {
+    name: 'login',
+    params: to?.params,
+    query: { redirectFrom: to.fullPath },
+    // hash: to?.hash,
+    replace: false,
+    force: true,
+  };
 }
-
-export { router };
