@@ -1,6 +1,7 @@
-import { store } from '@state/store';
-import { defineAsyncComponent, defineComponent, h } from 'vue';
-import { RouteRecordRaw } from 'vue-router';
+import { useAuthStore, useUserStore } from '@stores';
+import { defineAsyncComponent, defineComponent, h, type Component } from 'vue';
+import type { RouteRecordRaw } from 'vue-router';
+import { getLocaleFromParams, redirectToFallbackLocale } from '.';
 
 const routeRecordRaws: RouteRecordRaw[] = [
   {
@@ -10,26 +11,29 @@ const routeRecordRaws: RouteRecordRaw[] = [
      *
      */
     path: '/:locale?',
-    component: () => import(/* webpackPrefetch: true */ '@views/locale.vue'),
+    component: () => import('./views/locale.vue'),
     children: [
       {
         path: '',
         name: 'home',
-        component: () => lazyLoadView(import('@views/home.vue')),
+        component: () => lazyLoadView(import('./views/home.vue')),
       },
       {
         path: 'login',
         name: 'login',
-        component: () => lazyLoadView(import('@views/login.vue')),
+        component: () => lazyLoadView(import('./views/login.vue')),
         meta: {
-          beforeResolve: (to, from, next) => {
+          tmp: {},
+          beforeResolve: () => {
+            const authUserStore = useAuthStore();
+
             // If the user is already logged in
-            if (store.getters['auth/loggedIn']) {
+            if (authUserStore.loggedIn) {
               // Redirect to the home page instead
-              next({ name: 'home' });
+              return { name: 'home' };
             } else {
               // Continue to the login page
-              next();
+              return;
             }
           },
         },
@@ -37,40 +41,45 @@ const routeRecordRaws: RouteRecordRaw[] = [
       {
         path: 'profile',
         name: 'profile',
-        component: () => lazyLoadView(import('@views/profile.vue')),
+        component: () => lazyLoadView(import('./views/profile.vue')),
         meta: {
+          tmp: {},
           authRequired: true,
         },
-        props: () => ({ user: store.state.auth.currentUser || {} }),
+        props: () => {
+          const authUserStore = useAuthStore();
+
+          return {
+            user: authUserStore.currentUser || {},
+          };
+        },
       },
       {
         path: 'profile/:username',
         name: 'username-profile',
-        component: () => lazyLoadView(import('@views/profile.vue')),
+        component: () => lazyLoadView(import('./views/profile.vue')),
         meta: {
           authRequired: true,
           // HACK: In order to share data between the `beforeResolve` hook
           // and the `props` function, we must create an object for temporary
           // data only used during route resolution.
           tmp: {},
-          beforeResolve(to, from, next) {
-            store
+          beforeResolve: async (to) => {
+            try {
+              const userStore = useUserStore();
               // Try to fetch the user's information by their username
-              .dispatch('users/fetchUser', {
+              const userDetaul = await userStore.fetchUser({
                 username: to.params.username.toString(),
-              })
-              .then((user) => {
-                // Add the user to `meta.tmp`, so that it can
-                // be provided as a prop.
-                to.meta.tmp.user = user;
-                // Continue to the route.
-                next();
-              })
-              .catch(() => {
-                // If a user with the provided username could not be
-                // found, redirect to the 404 page.
-                next({ name: '404', params: { resource: 'User' } });
               });
+
+              // Add the user to `meta.tmp`, so that it can
+              // be provided as a prop.
+              to.meta.tmp.user = userDetaul;
+              // Continue to the route.
+              return true;
+            } catch (error: any) {
+              return { name: '404', query: { resource: 'User' } };
+            }
           },
         },
         // Set the user from the route params, once it's set in the
@@ -80,45 +89,53 @@ const routeRecordRaws: RouteRecordRaw[] = [
       {
         path: 'logout',
         name: 'logout',
-        component: () => lazyLoadView(import('@views/home.vue')),
+        component: () => lazyLoadView(import('./views/home.vue')),
         meta: {
+          tmp: {},
           authRequired: true,
-          beforeResolve(to, from, next) {
-            store.dispatch('auth/logOut');
+          beforeResolve(_, from) {
+            const authStore = useAuthStore();
+
+            authStore.logOut();
             const authRequiredOnPreviousRoute = from.meta.authRequired;
+
             // Navigate back to previous page, or home as a fallback
-            next(authRequiredOnPreviousRoute ? { name: 'home' } : { ...from });
+            return authRequiredOnPreviousRoute ? { name: 'home' } : { ...from };
           },
         },
       },
       {
         path: 'timeout',
         name: 'time-out',
-        component: () =>
-          lazyLoadView(
-            /* webpackPrefetch: true */ import('@views/_timeout.vue')
-          ),
+        component: () => lazyLoadView(import('./views/_timeout.vue')),
         meta: {
+          tmp: {},
           authRequired: false,
         },
       },
+      {
+        path: '404',
+        name: '404',
+        component: () => lazyLoadView(import('./views/_404.vue')),
+        // Allows props to be passed to the 404 page through route
+        // params, such as `resource` to define what wasn't found.
+        props: (route) => {
+          return { resource: route.query.resource };
+        },
+      },
     ],
-  },
-  {
-    path: '/404',
-    name: '404',
-    component: () =>
-      lazyLoadView(/* webpackPrefetch: true */ import('@views/_404.vue')),
-    // Allows props to be passed to the 404 page through route
-    // params, such as `resource` to define what wasn't found.
-    props: true,
   },
   // Redirect any unmatched routes to the 404 page. This may
   // require some server configuration to work in production:
   // https://router.vuejs.org/en/essentials/history-mode.html#example-server-configurations
   {
     path: '/:pathMatch(.*)*',
-    redirect: '404',
+    redirect: (to) => {
+      const localeInParams = getLocaleFromParams(to.params);
+
+      to.name = '404';
+      return redirectToFallbackLocale(localeInParams, to);
+    },
   },
 ];
 
@@ -127,35 +144,34 @@ const routeRecordRaws: RouteRecordRaw[] = [
 // back to a timeout view in case the page fails to load. You can
 // use this component to lazy-load a route with:
 //
-// component: () => lazyLoadView(import('@views/my-view'))
+// component: () => lazyLoadView(import('./views/my-view'))
 //
 // NOTE: Components loaded with this strategy DO NOT have access
 // to in-component guards, such as beforeRouteEnter,
 // beforeRouteUpdate, and beforeRouteLeave. You must either use
 // route-level guards instead or lazy-load the component directly:
 //
-// component: () => import('@views/my-view')
+// component: () => import('./views/my-view')
 //
-function lazyLoadView(AsyncView) {
+function lazyLoadView(AsyncView: Promise<Component>) {
   const AsyncHandler = defineAsyncComponent({
     loader: () => AsyncView,
     // A component to use while the component is loading.
-    loadingComponent: () => import('@views/_loading.vue'),
+    loadingComponent: () => import('./views/_loading.vue'),
     // Delay before showing the loading component.
     // Default: 200 (milliseconds).
     delay: 400,
     // A fallback component in case the timeout is exceeded
     // when loading the component.
-    errorComponent: () => import('@views/_timeout.vue'),
+    errorComponent: () => import('./views/_timeout.vue'),
     // Time before giving up trying to load the component.
     // Default: Infinity (milliseconds).
     timeout: 10000,
     /**
-     *
-     * @param {*} error Error message object
-     * @param {*} retry A function that indicating whether the async component should retry when the loader promise rejects
-     * @param {*} fail  End of failure
-     * @param {*} attempts Maximum allowed retries number
+     * @param error - Error message object
+     * @param retry - A function that indicating whether the async component should retry when the loader promise rejects
+     * @param fail - End of failure
+     * @param attempts - Maximum allowed retries number
      */
     onError(error, retry, fail, attempts) {
       if (error.message.match(/fetch/) && attempts <= 3) {
@@ -174,7 +190,7 @@ function lazyLoadView(AsyncView) {
       setup(props, { attrs, slots }) {
         return () => h(AsyncHandler, { ...attrs, ...props }, slots);
       },
-    })
+    }),
   );
 }
 
