@@ -1,10 +1,8 @@
-import { cy, describe, it } from 'local-cypress';
-import appConfig from '@src/app.config.json';
-
 describe('Home Page', () => {
   it('has the correct title and heading', () => {
     cy.visit('/');
-    cy.title().should('equal', `Home - ${appConfig.description}`);
-    cy.contains('h1', 'Home Page');
+    cy.title().should('equal', 'My App - Home');
+    cy.contains('h3', 'Home Page');
   });
 });
+export {};
