@@ -1,4 +1,4 @@
-import userHandlers from './routers/users';
 import authHandlers from './routers/auth';
+import userHandlers from './routers/users';
 
 export const handlers = [...authHandlers, ...userHandlers];
