@@ -1,16 +1,18 @@
+import Users from '@tests/mock-api/resources/users';
 import { rest } from 'msw';
-import { User } from '@models/user';
-import Users from '../resources/users';
 
 export default [
   /**
    * Log in a user with a username and password
    */
-  rest.post<User>('/api/session', async (req, res, ctx) => {
+  rest.post('/api/session', async (req, res, ctx) => {
     try {
-      const user = await Users.authenticate(req.body);
+      const body: { username: string; password: string } = await req.json();
+
+      const user = await Users.authenticate(body);
+
       return res(ctx.status(200), ctx.json(user));
-    } catch (error) {
+    } catch (error: any) {
       return res(ctx.status(401), ctx.json({ message: error.message }));
     }
   }),
@@ -21,15 +23,16 @@ export default [
   rest.get<any>('/api/session', (req, res, ctx) => {
     const currentUser = Users.findBy(
       'token',
-      req.headers.get('authorization') || ''
+      req.headers.get('authorization') || '',
     );
+
     if (!currentUser) {
       return res(
         ctx.status(401),
         ctx.json({
           message:
             'The token is either invalid or has expired. Please log in again.',
-        })
+        }),
       );
     }
     return res(ctx.status(200), ctx.json(currentUser));
@@ -38,7 +41,7 @@ export default [
   /**
    * A simple ping for checking online status
    */
-  rest.get('/api/ping', (req, res, ctx) => {
+  rest.head('/api/ping', (_req, res, ctx) => {
     return res(ctx.status(200), ctx.body('OK'));
   }),
 ];
