@@ -1,13 +1,14 @@
-import { rest } from 'msw';
-import Users from '../resources/users';
+import type { User } from '@src/models/user';
+import Users from '@tests/mock-api/resources/users';
+import { rest, type DefaultBodyType } from 'msw';
 
 export default [
-  rest.get<any, any, { username: string }>(
+  rest.get<any, { username: string }, User | DefaultBodyType>(
     '/api/users/:username',
     async (req, res, ctx) => {
       const currentUser = Users.findBy(
         'token',
-        req.headers.get('authorization') || ''
+        req.headers.get('authorization') || '',
       );
 
       if (!currentUser) {
@@ -16,21 +17,22 @@ export default [
           ctx.json({
             message:
               'The token is either invalid or has expired. Please log in again.',
-          })
+          }),
         );
       }
 
       const matchedUser = Users.findBy('username', req.params.username);
+
       if (!matchedUser) {
         return res(
           ctx.status(400),
           ctx.json({
             message: 'No user with this name was found.',
-          })
+          }),
         );
       }
 
       return res(ctx.status(200), ctx.json(matchedUser));
-    }
+    },
   ),
 ];
