@@ -2,6 +2,6 @@ import { setupServer } from 'msw/node';
 import { handlers } from './index';
 
 /**
- * Run on node environment. Like run with jest jsdom
+ * Run on node environment. Like run with vitest, jest jsdom
  */
 export const mockServer = setupServer(...handlers);
